pyproject.toml
[project]
name = "grokking"
version = "0.1.0"
description = "Toy examples of grokking for small transformers on arithmetic tasks."
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "antlr4-python3-runtime>=4.9.3",
    "certifi>=2025.1.31",
    "charset-normalizer>=3.4.1",
    "click>=8.1.8",
    "configparser>=7.2.0",
    "docker-pycreds>=0.4.0",
    "gitdb>=4.0.12",
    "gitpython>=3.1.44",
    "hydra-core>=1.3.2",
    "hydra-joblib-launcher>=1.2.0",
    "hydra-zen>=0.14.0",
    "idna>=3.10",
    "kaleido<=0.2",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "olefile>=0.47",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "promise>=2.3",
    "protobuf>=5.29.4",
    "psutil>=7.0.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "scikit-dimension>=0.3.4",
    "scikit-learn>=1.6.1",
    "sentry-sdk>=2.25.0",
    "shortuuid>=1.0.13",
    "six>=1.17.0",
    "smmap>=5.0.2",
    "subprocess32>=3.5.4",
    "termcolor>=2.3.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "transformers>=4.51.1",
    "typing-extensions>=4.13.0",
    "urllib3>=2.3.0",
    "wandb>=0.19.8",
    "yaspin>=3.1.0",
    "zarr>=3.0.6",
]

[project.scripts]
train_grokk = "grokking.scripts.train_grokk:main"

# Note: 
# The following section is needed on the HHU Hilbert HPC cluster to use the PyPI mirror,
# since we do not have internet access on the cluster.
# This should be removed when using the package outside of the cluster,
# and before the package is released.
#
# [[tool.uv.index]]
# name = "hhu"
# url = "http://pypi.repo.test.hhu.de/simple/"
# default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.ruff]
# extend = "$HOME/.config/ruff.toml"
target-version = "py312"
# Set the maximum line length.
line-length = 120

select = ["ALL"]
# Add the `line-too-long` rule to the enforced rule set. By default, Ruff omits rules that
# overlap with the use of a formatter, like Black, but we can override this behavior by
# explicitly adding the rule.
# extend-select = ["E501"]
ignore = ["RET504", "D102"]

[tool.ruff.lint.pylint]
max-args = 7

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pyright]
include = ["grokking", "tests"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "data",
    "hydra_output_dir",
    "wandb_output_dir",
]
ignore = ["src/oldstuff"]
defineConstant = { DEBUG = true }

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING"]
